Switched to update OpenGL uniform values in paintGL function instead. The reasons are listed in the comments in the code.
Set clear color to black. I had it at cyan for testing earlier.

Added reset view option to put view back in default location with
scaling and translation.

Added option to pan using middle mouse click with tool tip to show the
shortcut is space bar.

Added option to use mouse wheel to zoom in and out of the image. It
allows 5% to 300% scaling.

The application saves the previous folder opened, if it contained the
valid NIFTI images of course. This allows the application use this as
the starting path when opening the FileChooser dialog.

One thing to note is that when scaling or panning the image, the black
background can be noticeable depending on the colormap. I actually like
this feature because it is telling the user where the boundaries for the
drawing is at. Wherever there is black, this cannot be drawn on.

Added 'buddies' to labels so that pressing Alt+Key will give focus to
the label which in turn gives focus to its buddy. Keyboard shortcuts
were added for the four sliders and combo box in the settings group box.
The necessary shortcut is shown by the underlined letter in the label.
These are the labels and keyboard shortcuts, respectively:
Slice

Alt+S
Brightness

Alt+B
Contrast

Alt+C
Color Map

Alt+M

Added keyboard shortcut for open to Ctrl+O.