Fixed bug where increasing brightness wrapped the 1D texture color mapping around instead of clamping to the last value. So in the case of grayscale, instead of clamping values > 1.0 to a white value it would loop back around and make it black.
Added colormap images that are loaded on program startup and made into
1D textures. These 1D textures are used as look up tables to create the
color mapping effect

After a file is opened, the folder name is placed in the title bar so
you know what is open

Decremented all values in config.xml by one for the subject data because
the slice start/stop values were starting from 1 to the number of slices
and should be 0 to the number of slices. Basically, the values need to
be zero-indexed.

Something seemed fishy with some the subject data not being correctly
aligned. I went through and noticed that Subject 2 initial actually is
backwards as far as what the abdomen data meant. So instead of a lower
number meaning it was the upper slice, it meant it was the lower slice
and vice versa.

Remove status bar from GUI

Change getSlice in NIFTImage to getAxialSlice and add getCoronalSlice
and getSaggitalSlice.

Removed bit of code that uploaded MVP uniform to shader program on
initialization of OpenGL; however, this is redundant because it occurs
in paintGL() before being painted

Added CoronalSliceWidget which is attached to the smaller Widget below
the settings. It is sort of a "minimap" to show the user where they are
located for an axial slice. Currently, it is not possible to change
which coronal slice is used in the Widget. This will be fixed in future
releases.

The AxialSliceWidget and CoronalSliceWidget store a QVector4D which
represents the current location that the user is viewing. The X/W
components are not used but are provided so that it makes sense.