Clicking reset view applies to CoronalSliceWidget now
Added function initializeSliceView that initializeGL calls. The function
sets up the index buffer, vertex buffer, and other OpenGL related stuff.
Also, the function prints out version information regarding GPU and
OpenGL.

Renamed OpenGL items in AxialSliceWidget and CoronalSliceWidget to
include suffix slice if it a variable used to display the slice view.
With the addition of drawing lines or other stuff, the variable names
should be clear as to what they are meant for.

Fix bug where AxialSliceWidget and CoronalSliceWidget function isLoaded
returned inverted result. True if not loaded and false if loaded.

Add check for AxialSliceWidget and CoronalSliceWidget isLoaded() in
paintGL to prevent invalid memory access of fatImage/waterImage

Add cross hair line in CoronalSliceWidget to represent what axial slice
the user is viewing. The thickness of the crosshair line is based on the
scaling/translation of the coronal slice view. This means if the user
scales or translates the coronal slice view, the line will scale/
translate accordingly.

Add cross hair line in AxialSliceWidget to represent what coronal slice
the user is viewing. The thickness of the crosshair line is based on the
scaling/translation of the axial slice view. This means if the user
scales or translates the axial slice view, the line will scale/translate
accordingly.

Basic tracing functionality was added to AxialSliceWidget. A lot of
commented and ugly code is present because I was experimenting with
various methods of drawing lines. This should be fixed soon.