Rename curBrightness and curContrast to brightness and contrast, respectively. Unnecessary to say current when the old value isnt stored.
Rearranged AxialSliceWidget to have two color maps and two textures:
primary and secondary.

Added support for two additional SliceDisplayTypes: FatWater, WaterFat;
It is where one image is drawn directly over another with transparency
so that both can be seen at the same time. This is why the addition of a
primary/secondary distinction was necessary.

Added primTransparency and secdTransparency to AxialSliceWidget. This
includes implementing it to work in the shader and adding set/get
functions.

Renamed transparency to opacity because that is what A is in RGBA
format. That way I don't have to keep inverting it.