Added NumericType class to handle type conversion between OpenGL, OpenCV, and NIFTI.
Within the NIFTImage class, there was a few static functions used to
convert between the data type constants between the OpenGL, OpenCV, and
NIFTI libraries. Examples include DT_INT8 is the same as CV_8SC1 is the
same as GL_BYTE.

NumericType is a cleaner method of performing the same function. It
allows one to access a specified NumericType by doing
NumericType::OpenGL/OpenCV/NIFTI to get the NumericType pointer FOR the
specified parameters. Then, with the structure, one can get the other
corresponding types that are saved in the structure.

The structure is stored as a LUT that is hardcoded.