Added basic tracing functionality.
This includes a vector for each axial slice of a vector containing each
traceable fat layer containing a vector with the points for that layer
and axial slice.

The appropiate axial slice and each layer of points is drawn in paintGL.

Only layer 0 is drawn on currently and the GUI elements in the Tracing
tab do not work as of now.