Implemented appropiate transform operations to convert from NIFTI coords to OpenGL coords to Window coords.
I have made a decision that the coordinates stored in the vector are
NIFTI coordinates meaning that they are from (0 -> fatImage->getXDim() -
1, 0 -> fatImage->getYDim() - 1). See notebook for reasons why.

With that in mind, the coordinates need to be in window coordinates AND
must be scaled/translated appropiately with how AxialSliceWidget is
scaled. This involves matrix operations to transform the location
appropiately.