Added helper functions getXXXMatrix() to make point transformation much easier
Added getMVPMatrix, getNIFTIToOpenGLMatrix, getWindowToNIFTIMatrix, and
getWindowToOpenGLMatrix that returns the respective matrices to convert
from one coordinate system to another. There is an optional parameter in
the latter three functions that sets whether to include the MVP matrix
in the calculations.

Essentially, if you want the scaling and translation to be factored into
the matrix, then set this parameter to true.