Was doing MVP math backwards. It is projection * view * model
Inverted the y-axis on both slice widgets so that the origin is top left
like NIFTI, OpenCV, & Window coordinate system. This is typical in
applications.

Altered getXXX matrix functions to include a flipY parameter. This flips
the Y-axis. It is necessary in certain circumstances depending on if the
projectionMatrix has been multiplied through because it inverts the Y-
axis.

Cleaned up updateCrosshairLine to utilize matrix functions. This makes
the code much shorter and cleaner. It is likely more efficient as well
because you are doing one matrix calculation and using it for the
lineWidth and lineStart,lineEnd parameters.

Fixed bug where you could only draw from slice 0 -> getDim() - 2.
Before, you could not draw on the last slim getDim() - 1.

Flip NIFTI image in Y direction in addition to the Z direction now. This
is because the coronal slice numbers were not matching up with what is
in MATLAB. It was showing up okay in the viewer because the OpenGL
coordinate system is inverted (left-handed coordinates) compared to
MATLAB and many other viewers (right-handed coordinates).