Added default key shortcut for the file menu options. The Save As does not have a default on Windows and so I set one.
Added save and save as functionality to the buttons. Save will remember
the last path saved to and will continually save there. Save As asks
where to save. I am not pleased with the Save As overwrite dialog and
would like to work on being able to resize it.

Added additional checks to remove multiple points within the list.
Linear interpolation and the mouse press/release on same point caused
many duplicates to occur.

Added ability to import existing tracing data when the correct NIFTI
image is loaded.