Fix bug where mouse events before loading an image caused the application to crash. Check if loaded before doing anything in mouse event
Added version to application in About Dialog. This should also show up
in Windows properties for the EXE.

Also fix bug where warnings were not outputted on a release version.

Added the QT_MESSAGELOGCONTEXT preprocessor macro in the pro file when
building in release. This will ensure that the function and line is
reported in the message logger macros.