Added multiple views to the project.
Instead of one .UI file, there is now three. The main application which
has a dummy central widget within it. Then two widgets that are the
different views.

One view is meant for higher resolution monitors and has the coronal
minimap slice view split in half between the settings box. The other low
resolution view contains the coronal slice view minimap beside the
settings box.

This value is read on startup of application and the appropiate central
widget is started. I had to copy/paste the widget classes basically
twice to just tweak the UI.

This has the benefit of being able to make completely different views
and switch between them. However, the downside is that it is a lot of
extra code to manage and I'm not particularly sure this is the best way
to go about it.

Also, one downside is that the window view type is only checked on
startup of the application. It is possible to be able to switch views on
the fly, but it is going to take a bit of code to get all of the data
copied from one central widget to another. I need to reevaluate and see
if there are more efficient ways.