Added custom.pro which is included from VisceralFatValidation.pro
Added custom.pro to .gitignore

The purpose of the custom.pro is to contain any configuration values in
the pro file that are dependent on the user and should not be tracked in
GitHub. The most obvious is the external libraries because the path can
vary from person to person.