Moved fatImage, waterImage, tracingData, and subConfig to main window. These are the four pieces of data common to every view. They are passed down to each class.
Changed setImages function to setup in AxialSliceWidget/
CoronalSliceWidget and allow TracingData input for AxialSliceWidget.

Added isLoaded() boolean function to NIFTImage to determine whether
loaded. Instead of storing NULL pointer in the fatImage/waterImage
variables of AxialSliceWidget/CoronalSliceWidget, I am going to check
the class itself.

Added fattraces fragment and vertex shader that will be used to draw the
binary image for each fat layer

Implement a dirty integer into AxialSliceWidget and CoronalSliceWidget.
This uses Dirty enum to determine what needs to be updated on paint
(slice, crosshair, or traces)

One benefit of this is that supposedly calling update() repeatedly will
only result in one call to paintGL which means only once will the slice,
crosshair, and traces be updated. In addition, this is right before
drawing when it is needed most.

Made util::glCheckError a macro so that way the correct function name
and line is outputted when an OpenGL error occurs

Altered TracingPointsAddCommand to store a vector of points instead of
an index since that method is not used anymore. There is a new function
within it called addPoints that will actually compute the linear
interpolation and add any in between points. I REMOVED the smoothing
algorithm since I don't think it was doing anything before and it seems
fine now. If it becomes an issue, I can work on a new algorithm.

Updated fragment/vertex shaders to use up-to-date functions and not
deprecated. (textureXD is deprecated, varying/attribute deprecated, etc)

Added prompt when loading tracing data that will check if unsaved
tracing data already exists. If so, a prompt is given to confirm the
overwrite since old data will be discarded for the imported data.