Fixed bug where window title was not updated to display the directory when image is loaded
Changed Save and Save As to Save Tracing Data and Save Tracing Data As
to be more clear as to what is being saved.

Gave About button a shortcut of Ctrl+A.

Added Alt+X shortcuts for about and other menu items

Added shortcuts for visible tracing layers: Ctrl+1 to Ctrl+6 for the
layers. Updated tool tip to show shortcut.

Fix bug where location labels on status bar were added for each view
switch. Old location label is removed once view is deleted

The shader program log is only outputted if there is something to print,
such as warnings or errors compiling the program.

Record all data sent to/from console in a log file. Instead of
redirecting stdout to a file, that is left alone so that the messages
can be seen on the console AND in the log.