Instead of using an enum class for the Dirty bits which tells the widgets what textures to update, I switched to a namespace and constexpr of them so they can easily be ORed together.
Added check in loadTracingData to make sure the given point is within
the boundaries of the image first. Otherwise, OpenCV would throw an
Assert statement which kills the application.

Fix issue with using glTexSubImage2D to increase speed in tracing
layers. I had to have 7 booleans for each layer to know when they were
successfully allocated with just glTexImage2D.