Fix bug where exiting program caused crash because removing old status label was invalid since the parent window was already deleted.
Added some functions to TracingLayerData to make it easier to use. I
have added functions where you specify the x,y,z and it will set/reset
the data or get it. Makes it seem a bit nicer. My overall goal is to go
from using z, y, x to x, y, z. OpenCV has matrices backwards like that
and my method is to stay away from that and abstract it back to my way.