Added dirty bits for each of the tracing layers instead of just the current one.
When working with the command stack and quickly undoing changes on
multiple layers causes issues because the Dirty::Trace bit was set but
the appropiate texture was not updated in time before the current
tracing layer changed.

Instead, I added a separate bit for each of the tracing layers that will
be set when it needs to be redrawn.