Fixed bug where window coordinates were not correctly aligning with NIFTI coordinates. Removed update crosshair calculation beforehand and simply transform from NIFTI to window coordinates when drawing.
The bug was present with the crosshair line. If you went to the last y
slice of the image, it would not be all the way at the bottom of the
window which was odd. Next, when adding tracing points, sometimes the
placed tracing point was NOT where my mouse was hovering over. The
puzzling issue came down to the transformation matrices being scaled by
XDim and YDim of the NIFTI image when it should be XDim - 1 and YDim -
1. Zero-indexing makes that much of a difference...

Instead of doing the crosshair calculation and getting a lineStart,
lineEnd, and lineWidth (which has to be an integer width), I am going to
set the transformation matrix to be NIFTI->window and then simply draw
where it should be in the NIFTI coordinate system. This will make the
code easier to read and likely perform better since rounding errors will
be minimized. Removed updateCrosshair function and the variables
lineStart, lineEnd, and lineWidth.

Note: This means that the crosshair line will update on the fly and
there will be no need to call anything besides redrawing the screen.