Finished adding eraser functionality
Added a brush width combo box in the tracing tab. It is contained within
a stacked widget that is only set when the current draw mode is set to
DrawMode::Erase. There is a widget for each of the draw modes but this
is the only one used thusfar.

The widths can be 1px, 2px, 4px, 6px, or 8px. This is easily changed but
just arbitrarily picked numbers. A 'pixel' is defined as one NIFTI
coordinate point and not a window coordinate.