Save previous values of controls in settings.
This saves the previous values of the controls in the settings box. It
saves things such as brightness, contrast, primary color map, etc. This
way, if you have a preferred setting, you will not have to constantly
reset it over and over each time you load a NIFTI image.

This is done by propogating readSettings and writeSettings to the views
and slice widgets. There, each setting can be loaded and written.