Built VisceralFatValidation on Mac OS X. Fixed some things that caused warnings and errors on clang but not MSVC.
C++11 on clang requires a type-cast to integer when using in array
braces.

Another common error is -Wreorder which says that the initialization of
variables in the constructor is not in the same order as declared in the
header. I fixed a lot of these by simply moving where the variable was
declared in the header.

There was an error while compiling with clang with the VertexPT class
and undefined reference. Well, it turns out that it is NOT valid to
declare a constexpr in the header and then define it in a .cpp file.
Essentially, any constexpr is implicitly considered inline.

Apparently Apple does not support OpenGL compatibility profile.
Therefore, I had to make sure to request a Core profile context via
QSurfaceFormat::setProfile. See this link for specific details: http://
doc.qt.io/archives/qt-5.5/qabstractopenglfunctions.html

Do error checking for initializeOpenGLFunctions and report that the
version of OpenGL requested (in our case 3.3) was unable to be found.
Otherwise, it will crash the program and no reason why will be given.

Clang reports a few warnings with -Wmissing-braces and -Wreorder which I
have found to be unnecessary. As a result, I changed the .pro file to
ignore these.

There was an error where the menu wasn’t working on the Mac because the
application did not have focus on startup. You had to switch to another
application and then back to VisceralFatValidation to get the focus.
Removing the “config -= app_bundle” line from the .pro file did the
trick. I’m not exactly sure why this caused an error, but an app bundle
is likely what I want to do for OS X since it makes the entire folder
abstract into an icon and look nice.

Receiving error GL_INVALID_ENUM in glTexParameteri because it was meant
to be GL_CLAMP_TO_EDGE instead of GL_CLAMP? That’s weird it worked on
Windows.

Fixed bug where the axial slices were not displayed because the 2D
texture and 1D texture map were placed at the same active texture.
ALTHOUGH MANY THINGS are saying that this is allowed, I don’t think it
is! Read here: https://www.khronos.org/opengl/wiki/
Texture#Texture_image_units Read the note section! This is fine print
stuff that I missed. Moved mappingTexture to image unit 1 instead of 0.

One of the largest issues that I have had thus far is the fact that Mac
OS X does not support compatibility mode for OpenGL. It can only support
a core profile. The issue is that QPainter uses OpenGL ES 2.0 and is not
compatible with the 3.0 context. Compatibility mode fixes this and
allows both to be used but this is not so with core. Originally, I
attempted to convert my shaders and entire code to OpenGL ES 2.0 but I
received an error with the beginNativePainting function in QPainter. I’m
not sure why because that is not my code. This bug is fixed in “Add
support for OpenGL 3.2+ core profile contexts in QPainter.” It is
already merged and will be released with Qt 5.9. My current plan is to
try and build Qt 5.9 so I can get this functionality.

Added FPS counter to AxialSliceWidget for testing. This will be extended
in future commits.

Note: As of this commit, some functionality is commented out until Qt
5.9 is built. This is primarily the stuff that uses QPainter to render.