Created script to deploy mac applications
MacDeployScript runs macdeployqt to import Qt packages into the app
bundle.

Following that, the third-party libraries such as NIFTI and OpenCV are
manually added to the package and the libraries are changed so their
dependencies are relative to the app bundle.

Next, a DMG is created that includes the app bundle and a symlink to the
applications folder. The user upon installation simply moves the app
bundle to the applications symlink.

Finally, otool is ran on the libraries and application to allow the
developer to double-check that all dependencies are relative to the app
bundle now.