Rename VisceralFatValidation to SIUE-Fat-Segmentation-Tool. Whitespace is not allowed in the project name so the hyphens are used since this is also what GitHub does.
Increment version for renaming of application. As of v1.0.2 the application will be renamed! Update release date in About dialog.

Update windows installer for application rename

For windows installer, automatically run the windeployqt binary when compiling the script so that it will copy necessary Qt DLLs. The third-party DLLs are specified and copied over in the script as well.