Add auto-updater to program
Upon starting up the program, it will query Github for the latest
release and see if the user has the most up-to-date software. If not, it
prompts them to install it. On starting the program, updates are only
checked once a day, so if the user hits no then it will not bug them for
a day.

In additon, a check for updates button was added under the help menu in
case the user wants to check for updates manually. This supersedes the
'check once daily' parameter and will check for an update regardless.