Import QuaZip into project and switch tracing to load from a ZIP file instead of directory
QuaZip is a library that builds on the Qt framework to provide easy functionality to interacting with ZIP files. This includes features such as loading and saving files into the ZIP file, etc.

As a first pass, I got tracing converted to a zip file format with extension .sdt for SIUE Dixon Tracing.