Add Archiving Feature to NIFTI Subject Images
Instead of loading a directory with the NIFTI files and metadata contained inside, the new process is to load a .SDI file (SIUE Dixon Image). This makes it much easier to keep track of and no one will unintentionally edit one of the files within.

Note, in order for this to work with QuaZip, I had to create my own nifti_image_read_qt function. So, I imported the nifticlib to Qt Creator and named the project libNIFTI. There are a lot of small warnings and it really is hacky but it works for now.